requirements.txt
flask==3.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
requests==2.31.0
werkzeug==3.0.1
gunicorn==21.2.0
pillow==10.2.0
